Fix: narrow struct fields from ulong to uint32_t/os9addr_t
errortrap_typ: handleraddr/handlerstack → os9addr_t (68k entry points)
process_typ: memstart/memtop/icpta6/my_args → os9addr_t; cre_initsize → uint32_t
dir_typ: pos/fdcur → uint32_t (32-bit OS-9 file positions)
rbf_typ: all six ulong fields → uint32_t (device nr, positions, sector numbers)
file_rbf.c: replace pointer-arithmetic byte-count abuse with direct subtraction;
  Access_DirEntry deptr param → uint32_t*; remove unused bp variable
fileaccess.c: pDread pos local → uint32_t* to match dir_typ.pos

+++ Source/OS9exec_core/fileaccess.c
@@ -2401,9 +2401,9 @@ os9err pDread( _pid_, syspath_typ *spP, ulong *n, char* buffer )
 {
   os9err err;
     
-  ulong* pos  = &spP->u.disk.u.dir.pos; /* current file position */
-  ulong  offs = *pos & 0x1F;            /* offset    to start */
-  ushort index= *pos >> 5;              /* dir index to start */
+  uint32_t* pos  = &spP->u.disk.u.dir.pos; /* current file position */
+  uint32_t  offs = *pos & 0x1F;            /* offset    to start */
+  ushort    index= *pos >> 5;              /* dir index to start */
   char*  myb= buffer;
     
   ulong  cnt, nbytes;
